Applied pull requests from esp8266-oled-ssd1306: https://github.com/squix78/esp8266-oled-ssd1306/pull/169 https://github.com/squix78/esp8266-oled-ssd1306/pull/157 https://github.com/squix78/esp8266-oled-ssd1306/pull/144
And applied the suggestions from issues:
https://github.com/squix78/esp8266-oled-ssd1306/issues/170
https://github.com/squix78/esp8266-oled-ssd1306/issues/161

+++ lib/esp8266-oled-ssd1306/SSD1306Wire.h
@@ -80,7 +80,7 @@ class SSD1306Wire : public OLEDDisplay {
         // If the minBoundY wasn't updated
         // we can savely assume that buffer_back[pos] == buffer[pos]
         // holdes true for all values of pos
-        if (minBoundY == ~0) return;
+        if (minBoundY == (uint8_t)(~0)) return;
 
         sendCommand(COLUMNADDR);
         sendCommand(minBoundX);
